Rename register entries to define
