+++ docs/src/layouts/Root.astro
@@ -32,7 +32,7 @@ const { title, bodyClass } = Astro.props;
       })();
     </script>
     <script>
-      import "vrembem/register";
+      import "vrembem/define";
     </script>
   </head>
   <body class={bodyClass}>
