+++ packages/popover/vite.config.mjs
@@ -4,5 +4,5 @@ import data from "./package.json" with { type: "json" };
 
 export default createViteConfig(data, {
   index: resolve(import.meta.dirname, "./src/index.ts"),
-  register: resolve(import.meta.dirname, "./src/register.ts")
+  define: resolve(import.meta.dirname, "./src/define.ts")
 });
