+++ packages/vrembem/define.ts
@@ -1,0 +1,3 @@
+import "@vrembem/drawer/define";
+import "@vrembem/popover/define";
+export * from "./index.js";
