+++ packages/vrembem/package.json
@@ -21,9 +21,9 @@
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
     },
-    "./register": {
-      "types": "./dist/register.d.ts",
-      "import": "./dist/register.js"
+    "./define": {
+      "types": "./dist/define.d.ts",
+      "import": "./dist/define.js"
     },
     "./layers": {
       "sass": "./layers.scss",
@@ -59,7 +59,7 @@
     "dev",
     "dist",
     "index.ts",
-    "register.ts",
+    "define.ts",
     "index.scss",
     "layers.scss",
     "tokens.scss",
@@ -68,7 +68,7 @@
     "utilities.scss"
   ],
   "sideEffects": [
-    "./dist/register.js"
+    "./dist/define.js"
   ],
   "scripts": {
     "serve": "vite",
