--- packages/vrembem/register.ts
@@ -1,3 +1,0 @@
-import "@vrembem/drawer/register";
-import "@vrembem/popover/register";
-export * from "./index.js";
