--- vite.config.shared.d.ts
@@ -1,1 +1,0 @@
-export declare function createViteConfig(data: any, entry: any): import("vite").UserConfig;
